열거형 데이터 분리
