+++ src/main/java/view/InputResult.java
@@ -2,8 +2,7 @@ package view;
 
 import model.LottosData;
 import model.WinData;
-
-import static model.WinData.WinLottoCount.*;
+import model.WinLottoCount;
 
 public class InputResult {
     private final LottosData lottsData;
@@ -23,11 +22,11 @@ public class InputResult {
     public void goalResult() {
         System.out.println("\n당첨 통계");
         System.out.println("---");
-        System.out.println("3개 일치 (5,000원) -"+winData.getWinCountResult().get(threeCount.getCount())+"개");
-        System.out.println("4개 일치 (50,000원) -"+winData.getWinCountResult().get(fourCount.getCount())+"개");
-        System.out.println("5개 일치 (1,500,000원) -"+winData.getWinCountResult().get(fiveCount.getCount())+"개");
-        System.out.println("5개 일치, 보너스 볼 일치 (30,000,000원) -"+winData.getWinCountResult().get(fiveBonusCount.getCount())+"개");
-        System.out.println("6개 일치 (2,000,000,000원) -"+winData.getWinCountResult().get(sixCount.getCount())+"개");
+        System.out.println("3개 일치 (5,000원) -"+winData.getWinCountResult().get(WinLottoCount.threeCount.getCount())+"개");
+        System.out.println("4개 일치 (50,000원) -"+winData.getWinCountResult().get(WinLottoCount.fourCount.getCount())+"개");
+        System.out.println("5개 일치 (1,500,000원) -"+winData.getWinCountResult().get(WinLottoCount.fiveCount.getCount())+"개");
+        System.out.println("5개 일치, 보너스 볼 일치 (30,000,000원) -"+winData.getWinCountResult().get(WinLottoCount.fiveBonusCount.getCount())+"개");
+        System.out.println("6개 일치 (2,000,000,000원) -"+winData.getWinCountResult().get(WinLottoCount.sixCount.getCount())+"개");
     }
 
     public void profitRateResult(){
